+++ simpleor/cli.py
@@ -3,7 +3,7 @@ import sys
 import click
 import logging
 
-from simpleor import PROJECT_DIRECTORY
+from simpleor.utils import PROJECT_DIRECTORY
 from simpleor.monitoring import _configure_logger, LOGGING_LEVELS
 from simpleor.scheduler import (
     READ_OPTIONS,
