+++ simpleor/utils.py
@@ -1,0 +1,4 @@
+import os
+from pathlib import Path
+
+PROJECT_DIRECTORY = Path(os.path.dirname(os.path.realpath(__file__))).parent
